Tidying up the Tween Manager, ready for Timelines.
