+++ v3/src/tween/manager/components/Destroy.js
@@ -1,0 +1,7 @@
+//  Game level nuke
+var Destroy = function ()
+{
+    this.shutdown();
+};
+
+module.exports = Destroy;
