+++ v3/src/tween/manager/components/GetGlobalTimeScale.js
@@ -1,0 +1,6 @@
+var GetGlobalTimeScale = function ()
+{
+    return this.timeScale;
+};
+
+module.exports = GetGlobalTimeScale;
