+++ v3/src/tween/manager/components/ResumeAll.js
@@ -1,0 +1,13 @@
+var ResumeAll = function ()
+{
+    var list = this._active;
+
+    for (var i = 0; i < list.length; i++)
+    {
+        list[i].resume();
+    }
+
+    return this;
+};
+
+module.exports = ResumeAll;
