+++ v3/src/tween/manager/components/SetGlobalTimeScale.js
@@ -1,0 +1,8 @@
+var SetGlobalTimeScale = function (value)
+{
+    this.timeScale = value;
+
+    return this;
+};
+
+module.exports = SetGlobalTimeScale;
