+++ v3/src/tween/tween/components/CalcDuration.js
