+++ v3/src/tween/tween/components/Pause.js
