+++ v3/src/tween/tween/components/SetCallback.js
@@ -1,8 +1,8 @@
+var Tween = require('../Tween');
+
 var SetCallback = function (type, callback, params, scope)
 {
-    var types = [ 'onStart', 'onUpdate', 'onRepeat', 'onLoop', 'onComplete', 'onYoyo' ];
-
-    if (types.indexOf(type) !== -1)
+    if (Tween.TYPES.indexOf(type) !== -1)
     {
         this.callbacks[type] = { func: callback, scope: scope, params: params };
     }
